Promote record indexing to u64 across the workspace
Arena::record_count() called u32::try_from(records.len()).expect(...)
and would panic on a dataset exceeding 4.29B rows. The u32 ceiling was
also wired into RecordStorage / RecordView, every WindowContext
implementation's partition slice, the SecondaryIndex group lists,
HashAggregator's per-group input_rows + lineage, the sort utilities,
the executor's record_pos parameter, and the correlation-commit
orchestrator's seen_source_rows / RetractScope sets. Patching the call
site alone would leave the same ceiling baked in everywhere else.

Promote every record-index type to u64 in coordination so the ceiling
is removed, not papered over:
  - RecordStorage trait methods (resolve_field, resolve_qualified,
    available_fields, record_count) take/return u64.
  - RecordView holds u64; its FieldResolver impl follows.
  - GroupKeyError variants and value_to_group_key carry u64 row ids.
  - PartitionWindowContext.partition: &[u64], sum_slice positions: &[u64],
    sort_partition / is_sorted positions slices.
  - SecondaryIndex.groups maps to Vec<u64>; SecondaryIndex::get returns
    &[u64].
  - HashAggregator.input_rows: Vec<u64>, lineage: Vec<(u64, u32)> (group
    index stays u32 because it's a count of distinct groups, not record
    rows), retract_row(input_row_id: u64), input_rows_by_group_index
    returns Option<&[u64]>, GroupKey diagnostic carries row: u64.
  - Memory-accounting size_of calls follow the new shapes.
  - executor::evaluate_single_transform_windowed takes record_pos: u64;
    dispatch.rs feeds it via rn-1 / i as u64.
  - executor::commit::detect::RetractScope tracks BTreeSet<u64> and
    Vec<(NodeIndex, Vec<u64>)>; expand_with_dlq_events drops the
    u32::try_from(event.source_row) silent-truncation guard.
  - All bench fixtures (clinker-core/benches/{arena,sort,window}.rs,
    cxl/benches/eval.rs) and test doubles cascade.

Schema column indices (HashAggregator.group_by_indices) and per-group
indices (group_index, next_group_idx) remain u32 — they're not record
rows and would never approach the u32 ceiling.

cargo fmt, cargo clippy --workspace -- -D warnings, cargo test --workspace,
cargo check --benches --workspace, cargo check --features bench-alloc -p
clinker-benchmarks, and cargo test --benches -p clinker-benchmarks all
clean. cargo deny check is left to CI (not installed locally).

+++ crates/clinker-core/benches/window.rs
@@ -10,7 +10,7 @@ use std::io::{Cursor, Write};
 use std::sync::{Arc, Mutex};
 
 /// Build an Arena with numeric fields for window benchmarks.
-fn build_numeric_arena(partition_size: usize) -> (Arena, Vec<u32>) {
+fn build_numeric_arena(partition_size: usize) -> (Arena, Vec<u64>) {
     let schema = Arc::new(Schema::new(vec!["amount".into(), "category".into()]));
     let mut rng = fastrand::Rng::with_seed(42);
     let mut minimals = Vec::with_capacity(partition_size);
@@ -20,7 +20,7 @@ fn build_numeric_arena(partition_size: usize) -> (Arena, Vec<u32>) {
             Value::String(format!("cat_{}", i % 50).into_boxed_str()),
         ]));
     }
-    let positions: Vec<u32> = (0..partition_size as u32).collect();
+    let positions: Vec<u64> = (0..partition_size as u64).collect();
     (Arena::from_parts(schema, minimals), positions)
 }
